// src/app/models/tercero.model.ts
export class Tercero {
  id: number;
  siniestroId: number;
  nombres: string;
  rut: string;
  direccion: string;
  correo: string;
  correoAlt: string;
  telefono: string;
  telefonoAlt: string;
  alias: string;
  esPrincipal: Boolean;
  vehiculoPatente: string;
  vehiculoMarca: string;
  vehiculoModelo: string;
  vehiculoAnio: string;
  vehiculoColor: string;
  vehiculoNumeroMotor: string;
  vehiculoNumeroChasis: string;
  culpabilidad: string;
  clasificacion: string;
}